Merge of Torque2D 4.0 EA2

+++ engine/source/platform/types.posix.h
@@ -20,19 +20,22 @@
 // IN THE SOFTWARE.
 //-----------------------------------------------------------------------------
 
+#include <stddef.h>
+
 #ifndef _TYPESPOSIX_H_
 #define _TYPESPOSIX_H_
 
 
 #define FN_CDECL     ///< Calling convention
 
-// size_t is needed to overload new
-// size_t tends to be OS and compiler specific and may need to 
-// be if/def'ed in the future
-typedef unsigned int  dsize_t;      
+/**
+ * Platform-dependent type representing a size for pointers and arrays.
+ */
+
+typedef size_t dsize_t;      
 
 
-/** Platform dependent file date-time structure.  The defination of this structure
+/** Platform dependent file date-time structure.  The definition of this structure
   * will likely be different for each OS platform.
   */
 typedef S32 FileTime;
